モデルの作成を.py化
ファイルアップロードのtry, except
