+++ .vscode/settings.json
@@ -1,0 +1,3 @@
+{
+    "python.pythonPath": "/Users/keychi/.local/share/virtualenvs/image_ml-ruhc_TEP/bin/python"
+}
